+++ R/compute_stack.R
@@ -49,7 +49,7 @@ compute_stack.grouped_df <- function(x, stack_var = NULL, group_var = NULL) {
 
   x <- dplyr::ungroup(x)
   x <- compute_stack(x, stack_var, group_var)
-  x <- dplyr::group_by_(x, .dots = old_groups)
+  x <- dplyr::group_by(x, !!!old_groups)
 
   x
 }
@@ -59,28 +59,16 @@ compute_stack.data.frame <- function(x, stack_var = NULL, group_var = NULL) {
   assert_that(is.formula(stack_var), is.formula(group_var))
 
   # Round grouping variable to 8 significant digits
-  gvar <- lazyeval::interp(~round_fp(x), x = group_var[[2]])
-  x <- dplyr::group_by_(x, group__ = gvar)
+  x <- dplyr::group_by(x, group__ = round_fp(!!group_var[[2]]))
+
+  # Quiet R CMD check note
+  stack_upr_ <- NULL
 
   # FIXME: mutate evaluates in this function's environment, which isn't right.
-  # This is like mutate(x, stack_upr_ = cumsum(stack_var[[2]]),
-  #                     stack_lwr_ = lag(stack_upr_, default = 0))
-  # but with value of stack_var in the right place.
-
-  # dplyr 0.4.2 overrides lag instead of overriding default method AND
-  # preserves attributes so we don't want to use base lag
-  if (utils::packageVersion("dplyr") < "0.4.2") {
-    args <- list(
-      stack_upr_ = bquote(cumsum(.(stack_var[[2]]))),
-      stack_lwr_ = bquote(lag(stack_upr_, default = 0))
-    )
-  } else {
-    args <- list(
-      stack_upr_ = bquote(cumsum(.(stack_var[[2]]))),
-      stack_lwr_ = bquote(dplyr::lag(stack_upr_, default = 0))
-    )
-  }
-  x <- dplyr::mutate_(x, .dots = lazyeval::as.lazy_dots(args, asNamespace("ggvis")))
+  x <- dplyr::mutate(x,
+    stack_upr_ = cumsum(!!stack_var[[2]]),
+    stack_lwr_ = dplyr::lag(stack_upr_, default = 0)
+  )
 
   dplyr::ungroup(x)
 }